test-encdec: pass tags

+++ src/run-tests.sh
@@ -3,15 +3,22 @@
 # A shell script to run various tests automatically.
 #
 
+BAD_DIFF="DIFF!"
+BAD_VALGRIND="VALGRIND!"
+OK="OK"
+
+num_errs=0
+num_ok=0
+
 echo "stream-echo:"
 for testfile in ../tests/stream/*; do
 	outfile=$testfile.out
 	./test-stream $testfile $outfile
 
-	if ! diff $testfile $outfile; then
-		echo -e "\tFAILED $testfile"
+	if ! diff $testfile $outfile >/dev/null; then
+		echo -e "\t$BAD_DIFF $testfile"
 	else
-		echo -e "\tOK"
+		echo -e "\t$OK"
 	fi
 
 	rm -f $outfile
@@ -20,13 +27,27 @@ done
 echo "decenc:"
 for testfile in ../tests/libcbor/*.cbor; do
 	outfile=$testfile.out
-	./test-decenc $testfile $outfile
+	valgrind_result=$(valgrind ./test-decenc $testfile $outfile 2>&1 >/dev/null | tail -n1 | cut -d ' ' -f4,10)
 
-	if ! diff -q $testfile $outfile; then
-		echo -e "\tFAILED $testfile"
-	else
-		echo -e "\tOK"
+	err=0
+
+	if [ "$valgrind_result" != "0 0" ]; then
+		echo -e "\t$BAD_VALGRIND $testfile"
+		err=1
 	fi
 
-	rm -f $outfile
+	if ! diff -q $testfile $outfile >/dev/null; then
+		echo -e "\t$BAD_DIFF $testfile"
+		err=1
+	fi
+
+	if [ $err -eq 0 ]; then
+		echo -e "\tOK"
+		rm -f $outfile
+	fi
+
+	num_errs=$(expr $num_errs + $err)
 done
+
+echo
+echo "$num_errs error(s)";
